+++ haudiotagger/lib/src/rust/frb_generated.web.dart
@@ -10,6 +10,7 @@ import 'api/api.dart';
 import 'api/audio_properties.dart';
 import 'api/chapters.dart';
 import 'api/error.dart';
+import 'api/extended.dart';
 import 'api/normalization.dart';
 import 'api/picture.dart';
 import 'api/pipeline.dart';
@@ -55,6 +56,12 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   bool dco_decode_bool(dynamic raw);
 
   @protected
+  ExtendedChanges dco_decode_box_autoadd_extended_changes(dynamic raw);
+
+  @protected
+  ExtendedTag dco_decode_box_autoadd_extended_tag(dynamic raw);
+
+  @protected
   double dco_decode_box_autoadd_f_32(dynamic raw);
 
   @protected
@@ -92,6 +99,12 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   Chapter dco_decode_chapter(dynamic raw);
+
+  @protected
+  ExtendedChanges dco_decode_extended_changes(dynamic raw);
+
+  @protected
+  ExtendedTag dco_decode_extended_tag(dynamic raw);
 
   @protected
   double dco_decode_f_32(dynamic raw);
@@ -251,6 +264,13 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   bool sse_decode_bool(SseDeserializer deserializer);
 
   @protected
+  ExtendedChanges sse_decode_box_autoadd_extended_changes(
+      SseDeserializer deserializer);
+
+  @protected
+  ExtendedTag sse_decode_box_autoadd_extended_tag(SseDeserializer deserializer);
+
+  @protected
   double sse_decode_box_autoadd_f_32(SseDeserializer deserializer);
 
   @protected
@@ -292,6 +312,12 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   Chapter sse_decode_chapter(SseDeserializer deserializer);
+
+  @protected
+  ExtendedChanges sse_decode_extended_changes(SseDeserializer deserializer);
+
+  @protected
+  ExtendedTag sse_decode_extended_tag(SseDeserializer deserializer);
 
   @protected
   double sse_decode_f_32(SseDeserializer deserializer);
@@ -461,6 +487,14 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   void sse_encode_bool(bool self, SseSerializer serializer);
 
   @protected
+  void sse_encode_box_autoadd_extended_changes(
+      ExtendedChanges self, SseSerializer serializer);
+
+  @protected
+  void sse_encode_box_autoadd_extended_tag(
+      ExtendedTag self, SseSerializer serializer);
+
+  @protected
   void sse_encode_box_autoadd_f_32(double self, SseSerializer serializer);
 
   @protected
@@ -506,6 +540,13 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   void sse_encode_chapter(Chapter self, SseSerializer serializer);
+
+  @protected
+  void sse_encode_extended_changes(
+      ExtendedChanges self, SseSerializer serializer);
+
+  @protected
+  void sse_encode_extended_tag(ExtendedTag self, SseSerializer serializer);
 
   @protected
   void sse_encode_f_32(double self, SseSerializer serializer);